fieldfilters: split into new package
These don't work the same as our other filter types, so having everything under the
filters package doesn't really make sense. Split field filters out into their own package.
