+++ pkg/fieldfilters/fields_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Tetragon
 
-package filters
+package fieldfilters
 
 import (
 	"fmt"
